+++ docs/changelog.md
@@ -7,9 +7,13 @@ nav_order: 10
 # Changelog
 {: .no_toc }
 
+## [0.24.2]
+### Fixed
+- Table position meta data in millimeters.
+
 ## [0.24.1]
 ### Added
-- Table coordinates to measurement meta data.
+- Table position to measurement meta data.
 ### Fixed
 - Newline issue with CSV writer.
 ### Removed